feat: 🎸 use context provider solution
