+++ src/pages/alpha/TodoList/components/Form.jsx
@@ -1,7 +1,9 @@
 import { Input } from './Input';
 import Button from './Button';
+import { useTodoList } from './TodoListContext';
 
-export function Form({ onSubmit }) {
+export function Form() {
+  const { onSubmit } = useTodoList();
   return (
     <form
       className="flex w-full justify-center space-x-5 rounded-lg border-2 border-teal-200 bg-gray-100 p-8 shadow-lg"
